PR: Convert and split groupBy with dockey tests.
